% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{annotate_gse}
\alias{annotate_gse}
\title{Download and annotate a GSE study based on GSE number}
\usage{
annotate_gse(gse_id, method = "lv", label_method = "all", ...)
}
\arguments{
\item{gse_id}{GSE ID of the study to be annotated}

\item{method}{String similarity measure to be used. One of "jw", "jaro",
"sw", "nw", "me", "lv", "dl", "osa", "lcs", "qgram", "cosine", "jaccard"}

\item{label_method}{Labelling method to be used. One of "all", "common",
"unique", "lcsubstr", "lcsubseq".}

\item{...}{Other parameters to the \code{\link{annotate}} function}
}
\description{
Download and annotate a GSE study based on GSE number
}
