% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{annotateInterVar}
\alias{annotateInterVar}
\title{Annotate via InterVar's API.}
\usage{
annotateInterVar(vcf, liftover)
}
\arguments{
\item{vcf}{A variant table as the one created by `collectVars`}

\item{liftover}{Were variants lifted over?}
}
\value{
A nested list with InterVar's ACMG and other annotation pieces per variant.
}
\description{
Predict variants' pathogenicity according to
ACMG criteria as implemented by InterVar.
}
