% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\docType{package}
\name{annotator}
\alias{annotator}
\alias{annotator-package}
\title{Image annotation}
\description{
Image annotation using free drawing.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/valcu/annotator}
  \item Report bugs at \url{https://github.com/valcu/annotator/issues}
}

}
\author{
Mihai Valcu \email{mvalcu@gwdg.de}
}
\keyword{internal}
