% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_distance.R
\name{calculate_distance}
\alias{calculate_distance}
\title{Calculate distance}
\usage{
calculate_distance(annot, method, col, ...)
}
\arguments{
\item{annot}{A data.frame object containing sample annotation values}

\item{method}{String similarity measure to be used. One of ("jw", "jaro",
"sw1", "sw2", "nw1", "nw2", "nw3", "me", "lv", "dl", "osa",
"lcs", "qgram", "cosine", "jaccard")}

\item{col}{The column to be used to infer classes from. Default method "combine"
attempts to combine all columns for class inferral.}

\item{...}{Additional arguments to the distance functions}
}
\description{
Calculate distance
}
