% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organize.R
\name{clinvar_fetch}
\alias{clinvar_fetch}
\title{Fetch ClinVar variants.}
\usage{
clinvar_fetch(gns, path)
}
\arguments{
\item{gns}{Gene names character vector.}

\item{path}{Where data were downloaded}
}
\value{
vcf data.frame
}
\description{
Collect known variants for specific genes from ClinVar
}
