% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organize.R
\name{collectVars}
\alias{collectVars}
\title{Collect known variants.}
\usage{
collectVars(
  gns,
  databases = c("gnomad_man", "gnomad_auto", "clinvar", "lovd3", "all"),
  path,
  type = c("exomes", "genomes", "both"),
  liftover
)
}
\arguments{
\item{gns}{Gene names (HGNC) character vector.}

\item{databases}{Databases to retrieve variants from.}

\item{path}{Path to download directory.}

\item{type}{Type of gnomAD data to download. Used when `gnomad_auto` database is selected.}

\item{liftover}{Should variants be lifted over?}
}
\value{
A VCF-like data.frame with gene variants from the selected databases.
}
\description{
A wrapper function to collect known variants for specific genes
from public resources (gnomAD, ClinVar, LOVD3).
}
