% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organize.R
\name{constructAF}
\alias{constructAF}
\title{Retrieve gnomAD allele frequencies.}
\usage{
constructAF(
  gns,
  path,
  databases = c("gnomad_man", "gnomad_auto"),
  type = c("exomes", "genomes", "both"),
  liftover
)
}
\arguments{
\item{gns}{Gene names character vector.}

\item{path}{Where data were downloaded}

\item{type}{Type of gnomAD data to download.}

\item{liftover}{Should variants be lifted over?}
}
\value{
vcf data.frame
}
\description{
A wrapper function to retrieve population allele frequencies as recorded in gnomAD.
}
