% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organize.R
\name{constructAM}
\alias{constructAM}
\title{Contruct an AlphaMissense annotation file.}
\usage{
constructAM(gns, path, liftover)
}
\arguments{
\item{gns}{Gene names (HGNC) character vector.}

\item{path}{Path to download directory.}

\item{liftover}{Should variants be lifted over?}
}
\value{
A VCF-like data.frame with AlphaMissense pathogenicity prediction in the 3rd column (`ID`).
}
\description{
Download and organize AlphaMissense predictions for genes of interest.
}
