% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organize.R
\name{constructClinVar}
\alias{constructClinVar}
\title{Contruct ClinVar annotation file.}
\usage{
constructClinVar(gns, path)
}
\arguments{
\item{gns}{Gene names character vector.}

\item{path}{Where data were downloaded}
}
\value{
vcf data.frame
}
\description{
Download and organize ClinVar significance classification with
review status level for specific genes.
}
