% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organize.R
\name{constructClinVar_deprecated}
\alias{constructClinVar_deprecated}
\title{Contruct ClinVar annotation file.}
\usage{
constructClinVar_deprecated(gns, path, liftover)
}
\arguments{
\item{gns}{Gene names character vector.}

\item{path}{Where data were downloaded}

\item{liftover}{Should variants be lifted over?}
}
\value{
vcf data.frame
}
\description{
Download and organize ClinVar significance classification with
review status level for specific genes.
}
