% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_am}
\alias{download_am}
\title{Download AlphaMissense tsv files}
\usage{
download_am(path)
}
\arguments{
\item{path}{Where to download data}
}
\value{
AlphaMissense vcf files
}
\description{
This function downloads AlphaMissense tsv file.
}
