% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_clinvar}
\alias{download_clinvar}
\title{Download ClinVar files}
\usage{
download_clinvar(path)
}
\arguments{
\item{path}{Where to download data}
}
\value{
ClinVar vcf file.
}
\description{
This function downloads ClinVar vcf file.
}
