% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_lovd3}
\alias{download_lovd3}
\title{Download LOVD3 gene variants.}
\usage{
download_lovd3(gns, path)
}
\arguments{
\item{gns}{Gene names}

\item{path}{Where to download data}
}
\value{
LOVD3 gene files.
}
\description{
This function downloads LOVD3 gene.
}
