% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_revel}
\alias{download_revel}
\title{Download REVEL chromosome files.}
\usage{
download_revel(chr, path)
}
\arguments{
\item{chr}{Chromosome name}

\item{path}{Where to download data}
}
\value{
REVEL vcf files.
}
\description{
This function downloads REVEL vcf files per chromosome.
}
