% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_info_table.R
\name{get_info_table}
\alias{get_info_table}
\title{Get information table}
\usage{
get_info_table(check_dir, func, methods = c("jaro", "jw", "sw1", "nw1",
  "nw3", "me", "lv", "dl", "osa", "lcs", "qgram", "cosine", "jaccard"))
}
\arguments{
\item{check_dir}{folder containing studies}

\item{func}{a function for calculations. One of get_fnr, get_fpr or get_tpr.}

\item{methods}{the methods to evaluate. Default is all of ("jaro", "jw", "sw1", "nw1", 
"nw3", "me", "lv","dl", "osa", "lcs", "qgram", "cosine", "jaccard")}
}
\value{
a table containing the specified calculation for each study using each method
}
\description{
Get information table
}
