% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rand.R
\name{get_rand}
\alias{get_rand}
\title{Get Rand index of annotated samples}
\usage{
get_rand(check_dir, methods = c("jw", "sw1", "nw1", "me", "lv", "dl", "osa",
  "lcs", "qgram", "cosine", "jaccard"))
}
\arguments{
\item{check_dir}{Directory which to check the annotations}

\item{methods}{Similarity methods to get cluster values}
}
\description{
Get Rand index of annotated samples
}
