% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sample_annot.R
\name{get_sample_annot}
\alias{get_sample_annot}
\title{Get sample annotation data}
\usage{
get_sample_annot(out_file, geo_id)
}
\arguments{
\item{out_file}{File to write data to}

\item{geo_id}{GSE ID to get data}
}
\description{
Get sample annotation data
}
