% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpr_fpr.R
\name{get_tpr}
\alias{get_tpr}
\title{Calculate true positive rate between two clustering methods}
\usage{
get_tpr(group1, group2)
}
\arguments{
\item{group1}{The first clustering method}

\item{group2}{The reference ("true") method}
}
\description{
Calculate true positive rate between two clustering methods
}
\examples{
g1 <- sample(1:2, size=10, replace=TRUE)
g2 <- sample(1:3, size=10, replace=TRUE)
get_tpr(g1, g2)
}
