% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organize.R
\name{gnomad_fetch_manual}
\alias{gnomad_fetch_manual}
\title{Fetch manually downloaded gnomAD variants.}
\usage{
gnomad_fetch_manual(map, path, liftover, af = FALSE, sig = FALSE)
}
\arguments{
\item{map}{Gene data data.frame as returned from `map_fetch`.}

\item{path}{Path to download directory.}

\item{liftover}{Should variants be lifted over?}

\item{af}{Maintain allele frequencies?}
}
\value{
A VCF-like data.frame with gnomAD allele frequencies (af = TRUE) or ClinVar significance values (sig = TRUE) in the 3rd column (`ID`).
}
\description{
Collect known variants for specific genes downloaded manually from gnomAD
}
