% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdbscan2.R
\name{hdbscan2}
\alias{hdbscan2}
\title{Calculate HDBSCAN algorithm}
\usage{
hdbscan2(xdist, minPts = 5)
}
\arguments{
\item{xdist}{A distance matrix}

\item{minPts}{Minimum number of points per cluster}
}
\description{
This is a modified version of the hdbscan function from the dbscan package
in order to be able to calculate the KNN distance starting from a distance
matrix.
}
\keyword{internal}
