% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kNNdist2.R
\name{kNNdist2}
\alias{kNNdist2}
\title{Calculate k-Nearest Neighbors Distance}
\usage{
kNNdist2(sdm, k)
}
\arguments{
\item{sdm}{String Distance Matrix, as returned by the \code{calculate_distance} function}

\item{k}{Number of neighbors to compute distance}
}
\description{
Calculate k-Nearest Neighbors Distance
}
\keyword{internal}
