% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_annot.R
\name{label_annot}
\alias{label_annot}
\title{Automatic label creation}
\usage{
label_annot(annot0, method = "all")
}
\arguments{
\item{annot0}{Sample annotation data.frame with a column called "mega_col"
to infer labels from.}

\item{method}{Character string corresponding to a method to create the labels by.
One of "all", "common", "unique", "lcsubstr", "lcsubseq". See Details.}
}
\description{
Automatic label creation
}
\details{
The methods available for defining class labels are as follows:
\itemize{
   \item{all }{Return all labelling methods.}
   \item{common }{Strings shared by all instances of a class}
   \item{unique }{Strings unique to the instances of a particular class}
   \item{lcsubstr }{Longest common substring in a class}
   \item{lcsubseq }{Longest common subsequence in a class}
}
}
