% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{lift}
\alias{lift}
\title{LiftOver hg38 to hg19}
\usage{
lift(vcf, ...)
}
\arguments{
\item{vcf}{A vcf dataframe with CHR and POS columns}
}
\value{
Lifted over positions
}
\description{
LiftOver hg38 to hg19
}
