% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organize.R
\name{lovd3_fetch}
\alias{lovd3_fetch}
\title{Fetch lovd3 variants.}
\usage{
lovd3_fetch(gns, path)
}
\arguments{
\item{gns}{Gene names character vector.}

\item{path}{Where data were downloaded}
}
\value{
vcf data.frame
}
\description{
Collect known variants for specific genes from lovd3
}
