% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{map_fetch}
\alias{map_fetch}
\title{Annotate genes using Ensembl data}
\usage{
map_fetch(ensdb, gns, trans = FALSE)
}
\arguments{
\item{ensdb}{An EnsDb package.}

\item{gns}{Gene names (HGNC) character vector.}

\item{trans}{Should transcript level data be returned?}
}
\value{
A data frame with genes (default) or ensembl transcripts as rows and Ensembl annotation as columns.
}
\description{
Annotate genes using Ensembl data
}
