% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monge_elkan.R
\name{monge_elkan}
\alias{monge_elkan}
\title{Monge-Elkan string similarity}
\usage{
monge_elkan(A, B, split = " ", sim = "jw", p = 0.1)
}
\arguments{
\item{A}{String to be compared}

\item{B}{String to be compared}

\item{split}{Character to be used to split A and B (Default: " ")}

\item{sim}{Similatity method for intra-token distance (Default: "jw")}

\item{p}{Jaro-Winkler penalty (Default: 0.1)}
}
\description{
Monge-Elkan string similarity
}
