% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monge_elkan.R
\name{monge_elkan_matrix}
\alias{monge_elkan_matrix}
\title{Monge-Elkan string similarity matrix (w/ Jaro-Winkler intra-token distance)}
\usage{
monge_elkan_matrix(string_vecA, string_vecB)
}
\arguments{
\item{string_vecA}{Vector of strings to be compared}

\item{string_vecB}{Vector of strings to be compared}
}
\description{
Monge-Elkan string similarity matrix (w/ Jaro-Winkler intra-token distance)
}
