% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nw_metric.R
\name{nw_matrix}
\alias{nw_matrix}
\title{Needleman-Wunsch string similarity matrix}
\usage{
nw_matrix(string_vecA, string_vecB, ...)
}
\arguments{
\item{string_vecA}{Vector of strings to be compared}

\item{string_vecB}{Vector of strings to be compared}

\item{...}{Additional arguments to \code{nw_metrix()} function
(Default: NameNeedle::defaultNeedleParams)}
}
\description{
Needleman-Wunsch string similarity matrix
}
