% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pw_alignment.R
\name{pw_alignment_matrix}
\alias{pw_alignment_matrix}
\title{Pairwise alignment matrix}
\usage{
pw_alignment_matrix(string_vecA, string_vecB, type = "global",
  gapOpening = 10, gapExtension = 4, scoreOnly = TRUE)
}
\arguments{
\item{string_vecA}{Vector of strings to be compared}

\item{string_vecB}{Vector of strings to be compared}

\item{type}{Type of alignment}

\item{gapOpening}{Gap opening penalty}

\item{gapExtension}{Gap extension penalty}

\item{scoreOnly}{Return just the score for the similarity (Default: TRUE)}
}
\description{
Pairwise alignment matrix
}
