% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pw_alignment.R
\name{pw_alignment_metric}
\alias{pw_alignment_metric}
\title{Pairwise alignment}
\usage{
pw_alignment_metric(A, B, type = "global", gapOpening = 10,
  gapExtension = 4, scoreOnly = TRUE)
}
\arguments{
\item{A}{String to be compared}

\item{B}{String to be compared}

\item{type}{Type of alignment}

\item{gapOpening}{Gap opening penalty}

\item{gapExtension}{Gap extension penalty}

\item{scoreOnly}{Return just the score for the similarity (Default: TRUE)}
}
\description{
Pairwise alignment
}
