% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{revstatus_map}
\alias{revstatus_map}
\title{Review status on aggregate records (VCV and RCV) - ClinVar}
\format{
An object of class \code{data.frame} with 8 rows and 2 columns.
}
\usage{
revstatus_map
}
\description{
Review status on aggregate records (VCV and RCV) - ClinVar
}
\keyword{datasets}
