% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{runExample}
\alias{runExample}
\title{Run Shiny Examples}
\usage{
runExample(example, ...)
}
\arguments{
\item{example}{The name of the example to run. Available are: \verb{01_hello}, \verb{02_pipelineSimple}}

\item{...}{further arguments to pass to  \code{\link[shiny:runApp]{shiny::runApp()}}}
}
\description{
Launch Shiny example applications bundled in annotator.
}
\examples{
## Only run this example in interactive R sessions
if (interactive()) {
 annotator::runExample("01_hello")
 annotator::runExample("02_pipelineSimple")

 # Print the directory containing the code for all examples
 system.file("examples", package = "annotator")
}
}
