% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sw_metric.R
\name{sw_matrix}
\alias{sw_matrix}
\title{Smith-Waterman similarity matrix}
\usage{
sw_matrix(string_vecA, string_vecB)
}
\arguments{
\item{string_vecA}{Vector of strings to be compared}

\item{string_vecB}{Vector of strings to be compared}
}
\description{
Smith-Waterman similarity matrix
}
