% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sw_metric.R
\name{sw_metric}
\alias{sw_metric}
\title{Smith-Waterman similarity metric}
\usage{
sw_metric(A, B)
}
\arguments{
\item{A}{String to be compared}

\item{B}{String to be compared}
}
\description{
Smith-Waterman similarity metric
}
