% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{taste}
\alias{taste}
\title{Annotate via Mutation Taster's API.}
\usage{
taste(vcf)
}
\arguments{
\item{vcf}{A variant table as the one created by \code{collectVars}}
}
\value{
Annotated variants
}
\description{
Predict variants' pathogenicity using Mutation Taster.
}
