% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vcf_header_allele}
\alias{vcf_header_allele}
\title{VCF headers for IonReporter}
\format{
An object of class \code{character} of length 3.
}
\usage{
vcf_header_allele
}
\description{
VCF headers for IonReporter
}
\keyword{datasets}
