% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{annotate}
\alias{annotate}
\title{Create general purpose, IonReporter-compatible variant annotation files.}
\usage{
annotate(
  gns,
  annotators = c("revel", "alphamissense", "clinvar_sig", "gnomad_af", "intervar",
    "mutation_taster", "all"),
  databases = c("gnomad_man", "gnomad_auto", "clinvar", "lovd3", "all"),
  panelName = "./results/panel",
  path = "./dbs",
  type = c("exomes", "genomes", "both"),
  liftover = FALSE,
  saveRaw = FALSE
)
}
\arguments{
\item{gns}{Gene names (HGNC) character vector.}

\item{annotators}{Annotator resources to retrieve data from.}

\item{databases}{Databases to retrieve variants from. The argument is taken into consideration when `intervar`, `gnomad_af`, `mutation_taster` or `all` annotators are selected. `gnomad_man` expects manually downloaded gnomAD csv files placed in `gnomad/manual` subdirectory of the directory set in `path` argument (see below).}

\item{panelName}{Path and prefix for output file.}

\item{path}{Path to download directory.}

\item{type}{Type of gnomAD data to download. Required when `annotators` argument is set to/includes `gnomad_af`, `mutation_taster`, `intervar`, `all`}

\item{liftover}{Should variants be lifted over to hg19? (Default hg38)}

\item{saveRaw}{Save all variants without any annotation. Applicable only for `intervar` and `all` annotators.}
}
\value{
VCF-like annotation annotation files compatible with IonReporter.
}
\description{
A wrapper function for downloading and organizing annotation files.
}
