% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{annotate}
\alias{annotate}
\title{Create annotation files for IonReporter.}
\usage{
annotate(
  gns,
  annotators = c("revel", "alphamissense", "clinvar_sig", "intervar", "all"),
  databases = c("gnomad", "clinvar", "lovd3", "all"),
  panelName = "./results/panel",
  path = "./dbs",
  type = c("exomes", "genomes", "both"),
  saveRaw = FALSE
)
}
\arguments{
\item{gns}{Gene names character vector.}

\item{annotators}{Annotators to create files for.}

\item{panelName}{Path and prefix for output file.}

\item{path}{Where database files will be downloaded.}

\item{type}{Type of gnomAD data to download. Applicable only for \code{intervar} and \code{all} annotators.}

\item{saveRaw}{Save all variants without any annotation. Applicable only for \code{intervar} and \code{all} annotators.}
}
\value{
IonReporter annotation files.
}
\description{
A wrapper function for downloading and organizing annotation files.
}
