% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotator.R
\name{annotate}
\alias{annotate}
\title{Annotate an image
Annotate an image}
\usage{
annotate(im, resultId = "annot_id")
}
\arguments{
\item{im}{the input image. If missing a transparent 800x600 png is used.}

\item{resultId}{the id of the div in the UI where the annotation (the drawn polygon) is stored. Only relevant when the widget is used in shiny. Defaults to "annot_id".}
}
\description{
Annotate an image
Annotate an image
}
\examples{
if(interactive()) {
require(annotator)
im = system.file("sample_images", "aves","10.jpg" , package = "annotator")
annotate(im)
}
}
