% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotator.R
\name{annotate}
\alias{annotate}
\title{Annotate an image
Annotate an image}
\usage{
annotate(im, resultId = "annot_id")
}
\arguments{
\item{im}{input image}

\item{resultId}{the id  of the div in the UI where the annotation (the drawn polygon) is stored. Only relevant when the widget is used in shiny.}
}
\description{
Annotate an image
Annotate an image
}
\examples{
if(interactive()) {
require(annotator)
im = system.file("sample_images", "PUFFIN","010.jpg" , package = "annotator")
annotate(im)
}
}
