% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{annotateInterVar}
\alias{annotateInterVar}
\title{Annotate via InterVar's API.}
\usage{
annotateInterVar(vcf)
}
\arguments{
\item{vcf}{A variant table as the one created by \code{collectVars}}
}
\value{
Annotated variants
}
\description{
Predict variants' pathogenicity according to
ACMG criteria as implemented by InterVar.
}
