% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organize.R
\name{clinvar_fetch}
\alias{clinvar_fetch}
\title{Fetch ClinVar variants.}
\usage{
clinvar_fetch(gns, path, liftover)
}
\arguments{
\item{gns}{Gene names (HGNC) character vector.}

\item{path}{Path to download directory.}

\item{liftover}{Should variants be lifted over?}
}
\value{
A VCF-like data.frame with ClinVar variants.
}
\description{
Collect known variants for specific genes from ClinVar
}
