% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organize.R
\name{collectVars}
\alias{collectVars}
\title{Collect known variants.}
\usage{
collectVars(
  gns,
  databases = c("gnomad", "clinvar", "lovd3", "all"),
  path,
  type = c("exomes", "genomes", "both")
)
}
\arguments{
\item{gns}{Gene names character vector.}

\item{databases}{Databases to retrieve variants from.}

\item{path}{Where data were downloaded}

\item{type}{Type of gnomAD data to download.}
}
\value{
vcf data.frame
}
\description{
A wrapper function to collect known variants for specific genes
from public resources (gnomAD, ClinVar, LOVD3).
}
