% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organize.R
\name{collectVars}
\alias{collectVars}
\title{Collect known variants.}
\usage{
collectVars(gns, databases = c("gnomad", "clinvar", "lovd3", "all"), path)
}
\arguments{
\item{gns}{Gene names character vector.}

\item{databases}{Databases to retrieve variants from.}

\item{path}{Where data where downloaded}
}
\value{
vcf data.frame
}
\description{
Collect known variants for specific genes from public resources
(gnomAD, ClinVar, LOVD3).
}
