% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organize.R
\name{constructAF}
\alias{constructAF}
\title{Retrieve gnomAD allele frequencies.}
\usage{
constructAF(
  gns,
  path,
  databases = c("gnomad_man", "gnomad_auto"),
  type = c("exomes", "genomes", "both"),
  liftover
)
}
\arguments{
\item{gns}{Gene names (HGNC) character vector.}

\item{path}{Path to download directory.}

\item{type}{Type of gnomAD data to download. Used when `gnomad_auto` database is selected.}

\item{liftover}{Should variants be lifted over?}
}
\value{
A VCF-like data.frame with gnomAD allele frequency in the 3rd column (`ID`).
}
\description{
A wrapper function to retrieve population allele frequencies as recorded in gnomAD.
`gnomad_man` searches for manually downloaded gnomAD csv files in `gnomad/manual`
inside the `dbs` directory. If not found, it downloads (upon approval) gnomAD raw files instead.
}
