% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organize.R
\name{constructAM}
\alias{constructAM}
\title{Contruct AlphaMissense annotation file.}
\usage{
constructAM(gns, path, liftover)
}
\arguments{
\item{gns}{Gene names character vector.}

\item{path}{Where data were downloaded}

\item{liftover}{Should variants be lifted over?}
}
\value{
vcf data.frame
}
\description{
Download and organize AlphaMissense predictions for specific genes.
}
