% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organize.R
\name{constructClinVar}
\alias{constructClinVar}
\title{Retrieve ClinVar significance.}
\usage{
constructClinVar(gns, path, liftover)
}
\arguments{
\item{gns}{Gene names (HGNC) character vector.}

\item{path}{Path to download directory.}

\item{liftover}{Should variants be lifted over?}
}
\value{
A VCF-like data.frame with ClinVar germline classification values in the 3rd column (`ID`).
}
\description{
A wrapper function to retrieve ClinVar variant significance from gnomAD.
Manually downloaded gnomAD files are required.
}
