% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organize.R
\name{constructClinVar_deprecated}
\alias{constructClinVar_deprecated}
\title{Contruct ClinVar annotation file.}
\usage{
constructClinVar_deprecated(gns, path, liftover)
}
\arguments{
\item{gns}{Gene names (HGNC) character vector.}

\item{path}{Path to download directory.}

\item{liftover}{Should variants be lifted over?}
}
\value{
A VCF-like data.frame with ClinVar germline classification values and number of review stars in the 3rd column (`ID`).
}
\description{
Download and organize ClinVar significance classification with
review status level for specific genes.
}
