% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organize.R
\name{constructREVEL}
\alias{constructREVEL}
\title{Contruct REVEL annotation file.}
\usage{
constructREVEL(gns, path, liftover)
}
\arguments{
\item{gns}{Gene names (HGNC) character vector.}

\item{path}{Path to download directory.}

\item{liftover}{Should variants be lifted over?}
}
\value{
A VCF-like data.frame with REVEL deleteriousness prediction score in the 3rd column (`ID`).
}
\description{
Download and organize REVEL meta-predictions for specific genes.
}
