% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organize.R
\name{constructREVEL}
\alias{constructREVEL}
\title{Contruct REVEL annotation file.}
\usage{
constructREVEL(gns, path, liftover)
}
\arguments{
\item{gns}{Gene names character vector.}

\item{path}{Where data were downloaded}

\item{liftover}{Should variants be lifted over?}
}
\value{
vcf data.frame
}
\description{
Download and organize REVEL meta-predictions for specific genes.
}
