% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_gnomad}
\alias{download_gnomad}
\title{Download gnomad chromosome files}
\usage{
download_gnomad(chr, type = c("exomes", "genomes", "both"), path)
}
\arguments{
\item{chr}{Chromosome name}

\item{type}{Type of gnomAD data to download.}

\item{path}{Where to download data}
}
\value{
gnomAD vcf files
}
\description{
Downloads gnomAD vcf files according to specified chromosome.
}
