% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpr_fpr.R
\name{get_fnr}
\alias{get_fnr}
\title{Calculate false negative rate between two clustering methods}
\usage{
get_fnr(group1, group2)
}
\arguments{
\item{group1}{The first clustering method}

\item{group2}{The reference ("true") method}
}
\description{
Calculate false negative rate between two clustering methods
}
