% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organize.R
\name{gnomad_fetch_auto}
\alias{gnomad_fetch_auto}
\title{Fetch gnomAD variants.}
\usage{
gnomad_fetch_auto(
  map,
  type = c("exomes", "genomes", "both"),
  path,
  liftover,
  af = FALSE
)
}
\arguments{
\item{map}{Gene data as in map.}

\item{type}{Type of gnomAD data to download. Used when `gnomad_auto` database is selected.}

\item{path}{Path to download directory.}

\item{liftover}{Should variants be lifted over?}

\item{af}{Maintain allele frequencies?}
}
\value{
A VCF-like data.frame with gnomAD variants.
}
\description{
Collect known variants for specific genes from gnomAD.
}
