% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organize.R
\name{gnomad_fetch_manual}
\alias{gnomad_fetch_manual}
\title{Fetch manually downloaded gnomAD variants.}
\usage{
gnomad_fetch_manual(map, path, liftover, af = FALSE)
}
\arguments{
\item{map}{Gene data as in map.}

\item{path}{Where data were downloaded}

\item{liftover}{Should variants be lifted over?}

\item{af}{Maintain allele frequencies?}
}
\value{
vcf data.frame
}
\description{
Collect known variants for specific genes downloaded manually from gnomAD
}
