% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdbscan2.R
\name{hdbscan2}
\alias{hdbscan2}
\title{Calculate HDBSCAN algorithm}
\usage{
hdbscan2(xdist, minPts = 5, gen_hdbscan_tree = FALSE,
  gen_simplified_tree = FALSE)
}
\arguments{
\item{xdist}{A distance matrix}

\item{minPts}{Minimum number of points per cluster}

\item{gen_hdbscan_tree}{logical; should the robust single linkage tree be explicitly computed.}

\item{gen_simplified_tree}{logical; should the simplified hierarchy be explicitly computed.}
}
\description{
Calculate HDBSCAN algorithm
}
\keyword{internal}
