% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_annot.R
\name{label_annot}
\alias{label_annot}
\title{Automatic label creation}
\usage{
label_annot(annot0)
}
\arguments{
\item{annot0}{Sample annotation data.frame with a column called "mega_col"
to infer labels from.}
}
\description{
Automatic label creation
}
