% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{lift}
\alias{lift}
\title{LiftOver hg38 to hg19 genomic coordinates using rtracklayer functionalities}
\usage{
lift(vcf, ...)
}
\arguments{
\item{vcf}{A vcf dataframe with CHR and POS columns}

\item{...}{Arguments for methods}
}
\value{
A GRanges object of lifted over genomic coordinates
}
\description{
LiftOver hg38 to hg19 genomic coordinates using rtracklayer functionalities
}
