% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loop_studies.R
\name{loop_studies}
\alias{loop_studies}
\title{Loop over studies and annotate sample classes}
\usage{
loop_studies(studies_folder, out_folder, methods = c("jw", "sw", "nw",
  "me", "lv", "dl", "osa", "lcs", "qgram", "cosine", "jaccard"))
}
\arguments{
\item{studies_folder}{Directory from which to get study annotation files}

\item{out_folder}{Directory to write the annotated file to}

\item{methods}{Similarity measures to be applied to each sample annotation file}
}
\description{
Loop over studies and annotate sample classes
}
