% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organize.R
\name{lovd3_fetch}
\alias{lovd3_fetch}
\title{Fetch LOVD3 variants.}
\usage{
lovd3_fetch(gns, path, liftover)
}
\arguments{
\item{gns}{Gene names (HGNC) character vector.}

\item{path}{Path to download directory.}

\item{liftover}{Should variants be lifted over?}
}
\value{
A VCF-like data.frame with LOVD3 variants.
}
\description{
Collect known variants for specific genes from LOVD3
}
