% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{map_fetch}
\alias{map_fetch}
\title{Annotate genes using Ensembl data}
\usage{
map_fetch(ensdb, gns, trans = FALSE)
}
\arguments{
\item{ensdb}{An EnsDb package.}

\item{gns}{Gene names character vector.}

\item{trans}{Should return transcript data?}
}
\value{
Annotated genes
}
\description{
Annotate genes using Ensembl data
}
