% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nw_metric.R
\name{nw_metric}
\alias{nw_metric}
\title{Needleman-Wunsch string similarity}
\usage{
nw_metric(A, B, my_params)
}
\arguments{
\item{A}{String to be compared}

\item{B}{String to be compared}

\item{my_params}{Match, mismatch, gap and gap character values
(Default: NameNeedle::defaultNeedleParams)}
}
\description{
Needleman-Wunsch string similarity
}
