% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotator.R
\name{renderAnnotator}
\alias{renderAnnotator}
\title{Widget render function for use in Shiny}
\usage{
renderAnnotator(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{expr}{An annotator expression.}

\item{env}{A environment. Default to \code{parent.frame()}.}

\item{quoted}{A boolean value.}
}
\value{
An object generated by \code{\link[htmlwidgets:htmlwidgets-shiny]{htmlwidgets::shinyRenderWidget()}}.
}
\description{
Widget render function for use in Shiny
}
\examples{
# Print the directory containing the code for all examples to see this function in use.
system.file("examples", package = "annotator")
}
\seealso{
\code{\link[=runExample]{runExample()}}
}
